+++ RELEASENOTES.md
@@ -10,6 +10,8 @@
         still possible until the next major release using
         `setThrowsWhenUsingWrongThread(false)`
         ([#4463](https://github.com/google/ExoPlayer/issues/4463)).
+    *   Add a getter and callback for static metadata to the player
+        ([#7266](https://github.com/google/ExoPlayer/issues/7266)).
 *   Track selection:
     *   Add option to specify multiple preferred audio or text languages.
 *   Data sources: